requirements.txt
google-genai>=1.70.0
openai>=1.40.0
edge-tts>=6.1.0
moviepy>=2.0.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
pydub>=0.25.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
